Player visual script added (READ DESC)
I've written the script so that the player moves, before then performing an animation. The best way to go forward for interactions is probably for the player to trigger an event once it is done moving. you probably also want to add a vector3 as an "Interaction position", as in where the player moves when it interacts with an object. For the ground that would just be the position of the mouse.